% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFS.R
\name{autoFS}
\alias{autoFS}
\title{autoFS}
\usage{
autoFS(data_hex, x, y, num_of_model = 5, num_of_vi = 10)
}
\arguments{
\item{data_hex}{H2ODataFrame}

\item{x}{independent variables}

\item{y}{dependent variable}

\item{num_of_model}{the number of model to extract automl vi}

\item{num_of_vi}{the number of vi features to extract}
}
\description{
auto feature selection
}
\examples{
library(rAutoFS)
library(h2o)
h2o.init()
data(churn, package = "rAutoFS")
data_hex <- as.h2o(churn)
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
autoFS(data_hex, x, y, num_of_model=5, num_of_vi=10)
}
