% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automlVarImp.R
\name{automlVarImp}
\alias{automlVarImp}
\title{automlVarImp}
\usage{
automlVarImp(lb, num_of_model = 5, num_of_vi = 10)
}
\arguments{
\item{lb}{automl leaberboard}

\item{num_of_model}{the number of model to extract automl vi}

\item{num_of_vi}{the number of vi features to extract}
}
\description{
extract variable importance of automl
}
\examples{
library(rAutoFS)
library(h2o)
h2o.init()
data(churn, package = "rAutoFS")
data_hex <- as.h2o(churn)
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
splits <- h2o.splitFrame(data_hex, ratio = c(0.5, 0.3), seed = 1234)
train_hex <- splits[[1]]
valid_hex <- splits[[2]]
test_hex  <- splits[[3]]
aml <- h2o.automl(
  x = x, y = y,
  training_frame = h2o.rbind(train_hex, valid_hex),
  nfolds = 3,
  leaderboard_frame = test_hex,
  max_runtime_secs = 60*60,
  max_models = 60,
  exclude_algos = c("DeepLearning"),
  seed = 1234
)
lb <- aml@leaderboard
automlVarImp(lb=lb, num_of_model=5, num_of_vi=10)
}
